Move template stuff to Filesystem namespace and enable core extension
